Move populating changes view to onStart
This makes changes to the objects visible for validation when the modal
is "restarted" after tag editing.

Fixes: https://github.com/MarcusWolschon/osmeditor4android/issues/2291
